Atualização produção 08-12-2025

+++ back-end/WoopiAiHub.Application/Services/TeamServices.cs
@@ -6,6 +6,7 @@ using WoopiAiHub.Domain.DTOs.Response;
 using WoopiAiHub.Domain.Interfaces.Repository;
 using WoopiAiHub.Domain.Interfaces.Services;
 using WoopiAiHub.Domain.Models;
+using WoopiAiHub.Repository;
 
 namespace WoopiAiHub.Application.Services
 {
@@ -13,12 +14,21 @@ namespace WoopiAiHub.Application.Services
     {
         private readonly ITeamRepository _teamRepository;
         private readonly IUserRepository _userRepository;
+        private readonly IWorkflowServices _workflowServices;
+        private readonly IProfileRepository _profileRepository;
+        private readonly IWorkflowRepository _workflowRepository;
 
         public TeamServices(ITeamRepository teamRepository,
-                            IUserRepository userRepository)
+                            IUserRepository userRepository,
+                            IWorkflowRepository workflowRepository,
+                            IWorkflowServices workflowServices,
+                            IProfileRepository profileRepository)
         {
             _teamRepository = teamRepository;
             _userRepository = userRepository;
+            _workflowServices = workflowServices;
+            _profileRepository = profileRepository;
+            _workflowRepository = workflowRepository;
         }
 
         /// <summary>
@@ -100,9 +110,9 @@ namespace WoopiAiHub.Application.Services
         /// </summary>
         /// <param name="emailUser"></param>
         /// <returns></returns>
-        public async Task<ICollection<TeamDto>> FindByUser(string emailUser)
+        public async Task<ICollection<TeamDto>> FindAll()
         {
-            return await _teamRepository.FindAllByUser(emailUser).ToListAsync();
+            return await _teamRepository.FindAll().ToListAsync();
         }
 
         /// <summary>
@@ -134,11 +144,35 @@ namespace WoopiAiHub.Application.Services
                 }
             }
 
+            ICollection<Domain.Models.Profile> profiles = new List<Domain.Models.Profile>();
+            var profilesListCount = teamCreateDto.ProfileIds.Count() > 0;
+            if (profilesListCount)
+            {
+                team.Profiles.Clear();
+                profiles = _profileRepository.FindByIds(teamCreateDto.ProfileIds);
+
+                foreach(var profile in profiles)
+                {
+                    team.AddProfile(profile);
+                }
+            }
+
             var createResult = _teamRepository.CreateUniqueTeam(team);
             if (!createResult)
             {
                 throw new AppException(Domain.Enum.ErrorCode.Duplicated, "Duplicated Team Name", null);
             }
+
+            if (profilesListCount)
+            {
+                var workflows = await _workflowServices.FindByProfileStep(profiles);
+                foreach (var workflow in workflows)
+                {
+                    workflow.AddTeam(team);
+                }
+                await _workflowRepository.UpdateRange(workflows);
+            }
+
             return createResult;
         }
 
@@ -150,30 +184,63 @@ namespace WoopiAiHub.Application.Services
         /// <exception cref="ArgumentException"></exception>
         public async Task<bool> Update(TeamUpdateDto teamUpdateDto)
         {
-
             var team = _teamRepository.FindByIdReturnModel(teamUpdateDto.Id);
             if (team == null)
                 return false;
 
+            var teamsOldProfiles = team.Profiles.ToList();
             team.Update(teamUpdateDto.Name);
-
             if (teamUpdateDto.UserIds != null)
             {
                 team.Users.Clear();
                 var users = await _userRepository.FindByIdsAsync(teamUpdateDto.UserIds);
-
                 foreach (var user in users)
                 {
                     team.AddUser(user);
                 }
+
+                var updateUser = _teamRepository.Update(team);
+                if (!updateUser)
+                {
+                    throw new AppException(Domain.Enum.ErrorCode.DefaultError, "Error creating relationship to Users", null);
+                }
             }
 
-            var updateResult = _teamRepository.Update(team);
-            if (!updateResult)
+            ICollection<Domain.Models.Profile> profiles = new List<Domain.Models.Profile>();
+            var profilesListCount = teamUpdateDto.ProfileIds.Count() > 0;
+            team.Profiles.Clear();
+            if (profilesListCount)
+            {
+                profiles = _profileRepository.FindByIds(teamUpdateDto.ProfileIds);
+                foreach (var profile in profiles)
+                {
+                    team.AddProfile(profile);
+                }
+            }
+
+            var updateTeam = _teamRepository.Update(team);
+            if (!updateTeam)
             {
                 throw new AppException(Domain.Enum.ErrorCode.Duplicated, "Duplicated Team Name", null);
             }
-            return updateResult;
+
+            if (teamUpdateDto.UserIds != null)
+            {
+                var oldWorkflows = await _workflowServices.FindByProfileStep(teamsOldProfiles);
+                await _workflowServices.UpdateTeamWorkflowRelationship(team, oldWorkflows.ToList(), teamsOldProfiles);
+            }
+
+            if (profilesListCount)
+            {
+                var workflows = await _workflowServices.FindByProfileStep(profiles);
+                foreach (var workflow in workflows)
+                {
+                    workflow.AddTeam(team);
+                }
+                await _workflowRepository.UpdateRange(workflows);
+            }
+
+            return updateTeam;
         }
 
         /// <summary>
@@ -184,10 +251,21 @@ namespace WoopiAiHub.Application.Services
         public bool DeleteByIds(List<int> ids)
         {
             var teams = _teamRepository.FindByIds(ids);
-            bool hasDocuments = teams.Any(d => d.Documents.Count > 0);
+            var isAdmin = teams.Any(t => t.Name.Equals("admin", StringComparison.OrdinalIgnoreCase));
+            if (isAdmin)
+            {
+                throw new AppException(Domain.Enum.ErrorCode.InvalidValue, "Can't delete Admin team", null);
+            }
+
+            bool hasDocuments = teams.Any(d => d.Workflows.Count > 0);
             if (hasDocuments)
             {
                 throw new AppException(Domain.Enum.ErrorCode.InvalidValue, "Can't delete with documents", null);
+            }
+
+            foreach (var team in teams)
+            {
+                team.Profiles.Clear();
             }
             return _teamRepository.DeleteByIds(ids);
         }
